+++ CHANGELOG.md
@@ -10,6 +10,11 @@
     * ExtractUint16, ExtractUint32, ExtractUint64 opcodes
     * Txn, Global fields
 
+### Breaking changes
+
+`@algo-builder/runtime`:
++ Renamed `Runtime.getLogicSig` to `Runtime.createLsigAccount` #506.
+
 ### Bug Fixes
 +  Fix bug substring3 opcode pop wrong order [/#505](https://github.com/scale-it/algo-builder/pull/505), contribution: @vuvth.